fixed dtype deprecated warning

+++ lsd/compute_aic_all_peaks.py
@@ -194,9 +194,9 @@ def main():
     print('Elapsed time  = {:.2f} s'.format(end_time - start_time))
 
     # save AIC values
-    nib.Nifti1Image(aic_value.astype(np.float), affine).to_filename(args.oaic)
+    nib.Nifti1Image(aic_value.astype(np.float32), affine).to_filename(args.oaic)
     # save kernel MD
-    nib.Nifti1Image(kernel_MDs.astype(np.float), affine).to_filename(args.oMD)
+    nib.Nifti1Image(kernel_MDs.astype(np.float32), affine).to_filename(args.oMD)
 
 
 if __name__ == "__main__":
